MOSTRANDO DATOS FB

+++ app/src/main/java/edu/adf/profe/realtimedatabase/Cliente.kt
@@ -1,4 +1,4 @@
 package edu.adf.profe.realtimedatabase
 
 data class Cliente
-    (val edad: Int, val localidad:String, val nombre:String, val email:String, var clave:String="")
+    (val edad: Long, val localidad:String, val nombre:String, val email:String, var clave:String="")
